r124: * schroedinger/schroarith.c: * schroedinger/schroarith.h: * schroedinger/schrotables.c: * schroedinger/schrotables.h: * testsuite/arith_speed.c: * testsuite/tables.c: Update the arithmetic encoder to match the spec.  Changes BITSTREAM.  One of these days, I'll finally catch all the differences.

+++ schroedinger/schroedinger/schroarith.h
@@ -108,7 +108,7 @@ struct _SchroArith {
   uint8_t nextcode;
   int nextbits;
 
-  uint32_t *division_factor;
+  uint16_t *division_factor;
 };
 
 SchroArith * schro_arith_new (void);
